refactor: 코인 개수 자료구조 Map의 key변경
Coin enum클래스에서 Integer로 변경
